#21 refactor: create a database access object for user address data.
